refactor: remove the unused imports, change the name of files to meet the convention
